# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AgroTrack API",
    "description": "API para el sistema de gestión ganadera AgroTrack",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo AgroTrack",
      "email": "info@agrotrack.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Ganado": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "usuarioId": {
            "type": "integer",
            "format": "int64"
          },
          "identificacion": {
            "type": "string"
          },
          "raza": {
            "type": "string"
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date"
          },
          "sexo": {
            "type": "string",
            "enum": ["Macho", "Hembra"]
          },
          "peso": {
            "type": "number",
            "format": "float"
          },
          "estado": {
            "type": "string",
            "enum": ["Activo", "Vendido", "Fallecido"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/ganado": {
      "get": {
        "tags": ["Ganado"],
        "summary": "Obtener todos los registros de ganado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "raza",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por raza"
          },
          {
            "in": "query",
            "name": "sexo",
            "schema": {
              "type": "string",
              "enum": ["Macho", "Hembra"]
            },
            "description": "Filtrar por sexo"
          },
          {
            "in": "query",
            "name": "estado",
            "schema": {
              "type": "string",
              "enum": ["Activo", "Vendido", "Fallecido"]
            },
            "description": "Filtrar por estado"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Ordenar resultados (ej. peso:asc)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de registros de ganado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 2
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Ganado"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      },
      "post": {
        "tags": ["Ganado"],
        "summary": "Crear un nuevo registro de ganado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["identificacion", "raza", "fechaNacimiento", "sexo", "peso"],
                "properties": {
                  "identificacion": {
                    "type": "string"
                  },
                  "raza": {
                    "type": "string"
                  },
                  "fechaNacimiento": {
                    "type": "string",
                    "format": "date"
                  },
                  "sexo": {
                    "type": "string",
                    "enum": ["Macho", "Hembra"]
                  },
                  "peso": {
                    "type": "number",
                    "format": "float"
                  },
                  "estado": {
                    "type": "string",
                    "enum": ["Activo", "Vendido", "Fallecido"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registro de ganado creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Registro de ganado creado exitosamente"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Ganado"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ya existe un animal con esta identificación"
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      }
    },
    "/api/ganado/{id}": {
      "get": {
        "tags": ["Ganado"],
        "summary": "Obtener un registro de ganado por ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del registro de ganado"
          }
        ],
        "responses": {
          "200": {
            "description": "Registro de ganado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Ganado"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Registro no encontrado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      },
      "put": {
        "tags": ["Ganado"],
        "summary": "Actualizar un registro de ganado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del registro de ganado"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identificacion": {
                    "type": "string"
                  },
                  "raza": {
                    "type": "string"
                  },
                  "fechaNacimiento": {
                    "type": "string",
                    "format": "date"
                  },
                  "sexo": {
                    "type": "string",
                    "enum": ["Macho", "Hembra"]
                  },
                  "peso": {
                    "type": "number",
                    "format": "float"
                  },
                  "estado": {
                    "type": "string",
                    "enum": ["Activo", "Vendido", "Fallecido"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registro de ganado actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Registro de ganado actualizado exitosamente"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Ganado"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ya existe un animal con esta identificación"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Registro no encontrado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      },
      "delete": {
        "tags": ["Ganado"],
        "summary": "Eliminar un registro de ganado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del registro de ganado"
          }
        ],
        "responses": {
          "200": {
            "description": "Registro de ganado eliminado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Registro de ganado eliminado exitosamente"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Registro no encontrado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      }
    },
...
  }
}